Added Section Headers
WIP. There may need to be some restructuring as to where methods ends and where results starts. Typically a discussion is more the "so what?" of the paper and we would keep rote discussion of the results in the results but the rubric seems to request for a summary of results to be included in the discussion. It seems they want methods & results to be one section which is weird but iiwii. In any case, there's a good deal of reworking that I will get to tomorrow.

For anyone reading this wanting action items, we need to:
1) make everything more concise as we are well above the word limit
2) add figure numbers and a legend
3) flesh out the introduction better; perhaps move some content from the methods to the intro